GLState refactor - still not usable

+++ iOS/G3MiOSSDK/Commons/Mesh/DirectMesh.hpp
@@ -25,7 +25,7 @@ public:
              float lineWidth,
              Color* flatColor = NULL,
              IFloatBuffer* colors = NULL,
-             const float colorsIntensity = (float)0.0);
+             const float colorsIntensity = 0.0f);
 
   ~DirectMesh() {
